backend/requirements.txt
fastapi>=0.109,<1.0
uvicorn[standard]>=0.27,<1.0
pytest>=8.0,<9.0
pytest-cov>=4.1,<5.0
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
